% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration_helpers.R
\name{exponential_decay}
\alias{exponential_decay}
\title{helpers}
\usage{
exponential_decay(t, y0, k)
}
\value{
The return value, if any, from executing the utility.
}
\description{
utils function for infiltration analysis
}
