% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rainfall_analysis.R
\name{mod_rainfall_analysis_ui}
\alias{mod_rainfall_analysis_ui}
\title{rainfall_analysis UI Function}
\usage{
mod_rainfall_analysis_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
